backend/pyproject.toml
[tool.poetry]
name = "backend langchain"
version = "0.1.0"
description = ""
authors = ["Masfa Nasrullah Ansari"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain-core = "^0.3.12"
langchain-community = "^0.3.3"
langchain-google-genai = "^2.0.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
fastapi = "^0.115.3"
faiss-cpu = "1.8.0.post1"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"